% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{add_2OmeScore}
\alias{add_2OmeScore}
\title{Add 2-O-methylation score to PANORAMA object}
\usage{
add_2OmeScore(
  PANORAMA,
  lib_LowdNTPs,
  lib_HighdNTPs,
  newColName = "auto",
  flankSize = 5,
  perNuc_Zscaling = FALSE,
  onNucs = c("A", "C", "G", "T"),
  minCov = 50,
  nCores = 1
)
}
\arguments{
\item{PANORAMA}{list. PANORAMA object as output by \code{\link{panorama_PANORAMA}}}

\item{lib_LowdNTPs}{character. Name of group with low-dNTPs library preparation
found in PANORAMA$META$group}

\item{lib_HighdNTPs}{character. Name of group with high-dNTPs library preparation
found in PANORAMA$META$group}

\item{newColName}{character. Name of calculated metric to be stored in PANORAMA$RES,
assigned by default based on group_A and group_B}

\item{flankSize}{integer. Length of window to each side of position to
consider for metric calculation}

\item{onNucs}{character. Nucleotide(s) in which the metric will be calculated}

\item{minCov}{integer. Minimal coverage in position for metric to be calculated}
}
\description{
Methylation score that considers the log2-FCH of start ratio between
two RNA-seq libraries (low-dNTPs and high-dNTPs), subtracting the "noise"
from a surrounding window.
}
\details{
Incarnato et al. 2017: For Nm Incarnato and collabs. used the fold change
as the log2 of the ratio between the stoppage ratio at the low dNTPs
concentration and the stoppage ratio in the high dNTPs concentration sample.
Because the stoppage ratio seems to be strongly region specific, they
subtracted the local background defined as the mean ratio of the stoppage
in the neighborhood of a given position (+/- 5 nucleotides, excluding
given position).
}
