% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{add_biotype}
\alias{add_biotype}
\title{Add biotype (rRNA or tRNA)}
\usage{
add_biotype(txDT, geneInfo)
}
\arguments{
\item{txDT}{data.table}

\item{geneInfo}{data.frame Table with columns gene and biotype}
}
\value{
txDT
}
\description{
Adds RNA gene biotype based on gene name patterns
}
