% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{add_default_metrics}
\alias{add_default_metrics}
\title{Add all default metrics for Pan-Mod-seq}
\usage{
add_default_metrics(PANORAMA, nCores = 1)
}
\arguments{
\item{PANORAMA}{list. PANORAMA object as output by \code{\link{panorama_PANORAMA}}}

\item{nCores}{}
}
\description{
Add all default metrics for Pan-Mod-seq
}
