% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{add_default_metrics_v1}
\alias{add_default_metrics_v1}
\title{Add all default metrics for Pan-Mod-seq v1}
\usage{
add_default_metrics_v1(PANORAMA)
}
\arguments{
\item{PANORAMA}{list. PANORAMA object as output by \code{\link{panorama_PANORAMA}}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function was deprecated as the Pan-Mod-seq library treatments were expanded.
This functions applied only for the library treatments of lib 499.
}
