% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{add_ER1bpDS}
\alias{add_ER1bpDS}
\title{Add read-end rate 1bp downstream}
\usage{
add_ER1bpDS(
  PANORAMA,
  group_A,
  newColName = "auto",
  onNucs = c("A", "C", "G", "T"),
  minCov = 50,
  nCores = 1
)
}
\arguments{
\item{PANORAMA}{list. PANORAMA object as output by \code{\link{panorama_PANORAMA}}}

\item{group_A}{character. Name of group to be compared as found in PANORAMA$META$group}

\item{newColName}{character. Name of calculated metric to be stored in PANORAMA$RES,
assigned by default based on group_A and group_B}

\item{onNucs}{character. Nucleotide(s) in which the metric will be calculated}
}
\description{
Add read-end rate 1bp downstream
}
