% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{add_MRD}
\alias{add_MRD}
\title{Add Misincorporation Rate Difference}
\usage{
add_MRD(
  PANORAMA,
  group_A,
  group_B,
  newColName = "auto",
  onNucs = c("A", "C", "G", "T"),
  minReads = 30,
  nCores = 1
)
}
\arguments{
\item{PANORAMA}{list. PANORAMA object as output by \code{\link{panorama_PANORAMA}}}

\item{group_A}{character. Name of group to be compared as found in PANORAMA$META$group}

\item{group_B}{character. Name of group to be compared as found in PANORAMA$META$group}

\item{newColName}{character. Name of calculated metric to be stored in PANORAMA$RES,
assigned by default based on group_A and group_B}

\item{onNucs}{character. Nucleotide(s) in which the metric will be calculated}

\item{minReads}{numeric. Minimum number of nucleotides read in position to
compute the output value.}
}
\description{
Add Misincorporation Rate Difference
}
