% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{add_MRDns}
\alias{add_MRDns}
\title{Add Nucleotide Specific Misincorporation Rate Difference}
\usage{
add_MRDns(
  PANORAMA,
  group_A,
  group_B,
  refNuc,
  misNuc,
  minReads = 30,
  nCores = 1
)
}
\arguments{
\item{PANORAMA}{list. PANORAMA object as output by \code{\link{panorama_PANORAMA}}}

\item{group_A}{character. Name of group to be compared as found in PANORAMA$META$group}

\item{group_B}{character. Name of group to be compared as found in PANORAMA$META$group}

\item{refNuc}{character. Reference nucleotide}

\item{misNuc}{character. Misincorporated nucleotide}

\item{minReads}{numeric. Minimum number of nucleotides read in position to
compute the output value.}
}
\description{
Add Nucleotide Specific Misincorporation Rate Difference
}
