% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{add_scoreZ_3p}
\alias{add_scoreZ_3p}
\title{Add scoreZ accounting for read-ends (3prime)}
\usage{
add_scoreZ_3p(
  PANORAMA,
  group_A,
  newColName = "auto",
  onNucs = c("A", "C", "G", "T"),
  minReads = 30,
  flankSize = 6,
  nCores = 1
)
}
\arguments{
\item{PANORAMA}{list. PANORAMA object as output by \code{\link{panorama_PANORAMA}}}

\item{group_A}{character. Name of group to be compared as found in PANORAMA$META$group}

\item{newColName}{character. Name of calculated metric to be stored in PANORAMA$RES,
assigned by default based on group_A and group_B}

\item{onNucs}{character. Nucleotide(s) in which the metric will be calculated}

\item{minReads}{numeric. Minimum number of nucleotides read in position to
compute the output value.}

\item{flankSize}{integer. Length of window to each side of position to
consider for metric calculation}
}
\description{
Add scoreZ accounting for read-ends (3prime)
}
