% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{addKnownRNAmods}
\alias{addKnownRNAmods}
\title{Add known RNAmods to PANORAMA object}
\usage{
addKnownRNAmods(PANORAMA, RNAmods)
}
\arguments{
\item{RNAmods}{}
}
\description{
Adds a column 'nuc' to RES and CALLS, using the column 'pos' as reference
Reference RNAmod annotation should have at least two columns:
\itemize{
\item 'pos': Position given by 'gene:txcoor'
\item 'nuc': RNA modification's short name.
}
}
