% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{call_RNAmods_linRegMods}
\alias{call_RNAmods_linRegMods}
\title{Call RNA modifications using linear regression models}
\usage{
call_RNAmods_linRegMods(PANORAMA, models, nucBias = TRUE)
}
\arguments{
\item{nucBias}{}
}
\value{
PANORAMA object
}
\description{
Call RNA modifications using linear regression models
}
