% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{call_RNAmods_logRegMods}
\alias{call_RNAmods_logRegMods}
\title{Logistic regressions scores}
\usage{
call_RNAmods_logRegMods(PANORAMA, logMods, nucBias = TRUE)
}
\arguments{
\item{logMods}{}
}
\description{
Update PANORAMA$CALLS with logistic and linear scores and prediction
}
