% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_whichRNAmods}
\alias{check_whichRNAmods}
\title{Check RNAmods}
\usage{
check_whichRNAmods(PANORAMA)
}
\arguments{
\item{PANORAMA}{}
}
\value{
character
}
\description{
Check which RNAmods are possible to predict based on RNAmod metrics present
in PANORAMA object
}
