% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{confRNAmods_list}
\alias{confRNAmods_list}
\title{List of confounding RNAmods}
\format{
list
}
\usage{
confRNAmods_list
}
\description{
A list containing the confounding RNAmods per RNAmod identified by PANORAMA
}
\keyword{datasets}
