% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuebox.R
\name{duree_moyenne}
\alias{duree_moyenne}
\title{Durée moyenne des projets (en années)}
\usage{
duree_moyenne(table)
}
\arguments{
\item{table}{a data frame}
}
\value{
a formatted number
}
\description{
Durée moyenne des projets (en années)
}
\examples{
duree_moyenne(table = table_panorama)

}
