% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuebox.R
\name{euros}
\alias{euros}
\title{Total cost of all projects}
\usage{
euros(table)
}
\arguments{
\item{table}{}
}
\value{
a string with the total cost of the projects
}
\description{
Total cost of all projects
}
\examples{
euros(table = table_panorama)
}
