% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{hlp_assign_scores}
\alias{hlp_assign_scores}
\title{Assign scores to respective RNAmods in PANORAMA$CALLS}
\usage{
hlp_assign_scores(PANORAMA, RNAmod, scores, onNucs = c("A", "C", "G", "T"))
}
\arguments{
\item{PANORAMA}{list. PANORAMA object as output by \code{\link{panorama_PANORAMA}}.}

\item{RNAmod}{character. Name of RNAmod to assign scores to.}

\item{scores}{character. Names of scores to be assigned to RNAmod.}

\item{onNucs}{character. Vector of base nucleotides to include in scores, all
nucleotides by default, i.e. c("A", "C", "G", "T").}
}
\description{
Assign scores to respective RNAmods in PANORAMA$CALLS
}
