% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{hlp_start_CALLS}
\alias{hlp_start_CALLS}
\title{Initialize PANORAMA$CALLS}
\usage{
hlp_start_CALLS(PANORAMA)
}
\arguments{
\item{PANORAMA}{list. PANORAMA object as output by \code{\link{panorama_PANORAMA}}}
}
\description{
Initialize PANORAMA$CALLS
}
