% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_table.R
\name{import_table}
\alias{import_table}
\title{Import the data into R}
\usage{
import_table(path)
}
\arguments{
\item{path}{the path to the data}
}
\value{
a tibble
}
\description{
Import the data into R
}
\examples{
table_panorama <- import_table(path = "raw-data/panorama_projets_si.csv", n = 56)

}
