% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuebox.R
\name{interministeriel}
\alias{interministeriel}
\title{Nombre de projets interministériels}
\usage{
interministeriel(table)
}
\arguments{
\item{table}{}
}
\value{
a number
}
\description{
Nombre de projets interministériels
}
\examples{
interministeriel(table = table_panorama)
}
