% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_on_Docker.R
\name{libComplexReport_docker}
\alias{libComplexReport_docker}
\title{Library complexity report}
\usage{
libComplexReport_docker(
  META,
  maxExtrapolation = 2010000,
  steps = 1e+05,
  verbose = FALSE
)
}
\arguments{
\item{verbose}{}
}
\description{
Library complexity report
}
