% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable.R
\name{make_html_table}
\alias{make_html_table}
\title{Make an HTML table}
\usage{
make_html_table(table)
}
\arguments{
\item{table}{the name of the table to be used}
}
\value{
an html table
}
\description{
Make an HTML table
}
\examples{
make_html_table(table = table_panorama)
}
