% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{maxMCC}
\alias{maxMCC}
\title{Maximum MCC}
\usage{
maxMCC(scores, successes, ncores = 1L)
}
\arguments{
\item{scores}{numeric}

\item{successes}{logical}

\item{ncores}{integer}
}
\value{
numeric
}
\description{
Maximum MCC
}
