% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{MCC}
\alias{MCC}
\title{Matthews Correlation COefficient}
\usage{
MCC(scores, successes, thr)
}
\arguments{
\item{scores}{numeric}

\item{successes}{logical}

\item{thr}{numeric}
}
\value{
numeric
}
\description{
Matthews Correlation COefficient
}
