% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{miniMETA}
\alias{miniMETA}
\title{Mini-Meta (with ordered treatments)}
\usage{
miniMETA(PANORAMA)
}
\arguments{
\item{PANORAMA}{list. PANORAMA object}
}
\description{
Smaller version of experimental design for specific purposes
}
