% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_on_Docker.R
\name{mkSTARgenome_docker}
\alias{mkSTARgenome_docker}
\title{Make STAR genome}
\usage{
mkSTARgenome_docker(
  fastaGenome,
  bedAnnotation = NULL,
  outDir = NULL,
  nCores = 2,
  maxReadLength = 101
)
}
\arguments{
\item{maxReadLength}{}
}
\description{
Wrapper for STAR genome generation
}
