% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{modelSummary}
\alias{modelSummary}
\title{Models summary table}
\usage{
modelSummary(models)
}
\arguments{
\item{models}{}
}
\value{
data.table
}
\description{
Generates a table summarizing the models variables, coefficients and
correlation to the fitted results used in training.
}
