% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuebox.R
\name{n_projects}
\alias{n_projects}
\title{Count the number of projects}
\usage{
n_projects(table)
}
\arguments{
\item{table}{a data frame}
}
\value{
a scalar
}
\description{
Count the number of projects
}
\examples{
n_projects(table = table_panorama)
}
