% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{palette_sgmap}
\alias{palette_sgmap}
\title{Color palette with SGMAP's colors}
\usage{
palette_sgmap()
}
\value{
a palette
}
\description{
Color palette with SGMAP's colors
}
\examples{
palette_sgmap()

}
