% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PANORAMA-package.R
\docType{package}
\name{PANORAMA}
\alias{PANORAMA}
\alias{PANORAMA-package}
\title{PANORAMA: A package to process Pan-Mod-seq data}
\description{
A package to process Pan-Mod-seq data
}
\author{
\strong{Maintainer}: Miguel Angel Garcia-Campos \email{garciacampos.bioinfo@gmail.com} (\href{https://orcid.org/0000-0001-5223-8174}{ORCID}) (https://angelcampos.github.io/)

}
\keyword{Epitranscriptomics}
\keyword{RNASeq}
