% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{panorama_add_scoreA5p}
\alias{panorama_add_scoreA5p}
\title{Add scoreA 5prime}
\usage{
panorama_add_scoreA5p(
  PANORAMA,
  group_A,
  newColName = "auto",
  onNucs = c("A", "C", "G", "T"),
  minMedCov = 15
)
}
\arguments{
\item{minMedCov}{}
}
\description{
Add scoreA 5prime
}
