% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{panorama_assignScores}
\alias{panorama_assignScores}
\title{Assigning default Scores to tables by sets of samples}
\usage{
panorama_assignScores(PANORAMA)
}
\arguments{
\item{PANORAMA}{list. PANORAMA object as output by \code{\link{panorama_PANORAMA}}}
}
\description{
Assigning default Scores to tables by sets of samples
}
