% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{panorama_baseCols}
\alias{panorama_baseCols}
\title{PANORAMA base columns}
\format{
character
}
\usage{
panorama_baseCols
}
\description{
Used internally by PANORAMA but may be useful
}
\keyword{datasets}
