% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{panorama_baseCoorCols}
\alias{panorama_baseCoorCols}
\title{PANORAMA base coordinate columns}
\format{
character
}
\usage{
panorama_baseCoorCols
}
\description{
Used internally by PANORAMA but may be useful
}
\keyword{datasets}
