% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{panorama_confMatrix}
\alias{panorama_confMatrix}
\title{Confusion matrix}
\usage{
panorama_confMatrix(
  PANORAMA,
  nucBias = TRUE,
  removeOtherRNAmods = TRUE,
  rmNotModGenes = TRUE,
  blacklist = NULL
)
}
\arguments{
\item{blacklist}{}
}
\value{
data.table
}
\description{
Confusion matrix generation for known sites in PANORAMA object. Known sites are
added with
}
