% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{panorama_insTable}
\alias{panorama_insTable}
\title{Generate inspection table from PANORAMA object}
\usage{
panorama_insTable(PANORAMA, nCores)
}
\arguments{
\item{PANORAMA}{list. PANORAMA object}

\item{nCores}{numeric. Number of cores used for multi-thread processes}
}
\value{
list containing inspection table and sets vector
}
\description{
Generate inspection table from PANORAMA object
}
