% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{panorama_makeCalls}
\alias{panorama_makeCalls}
\title{Make calls (deprecated)}
\usage{
panorama_makeCalls(PANORAMA)
}
\arguments{
\item{PANORAMA}{list. PANORAMA object}
}
\description{
Make calls (deprecated)
}
