% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{panorama_metricsBoxPlot_byNuc}
\alias{panorama_metricsBoxPlot_byNuc}
\title{Boxplots of $RES calculated metrics grouping by modified nucleotides}
\usage{
panorama_metricsBoxPlot_byNuc(PANORAMA, metrics = NULL, title = "")
}
\arguments{
\item{PANORAMA}{list. PANORAMA object as output by \code{\link{panorama_PANORAMA}}}

\item{metrics}{character. Selected metrics to plot. By default plots all metrics}

\item{title}{character. Title to be shown on top of plot}
}
\description{
Boxplots of $RES calculated metrics grouping by modified nucleotides
}
