% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{panorama_metricsScatterPlot_byPos}
\alias{panorama_metricsScatterPlot_byPos}
\title{PANORAMA scatterplot of scores by txcoor}
\usage{
panorama_metricsScatterPlot_byPos(
  PANORAMA,
  metrics = NULL,
  title = "",
  na.rm = FALSE
)
}
\arguments{
\item{PANORAMA}{list. PANORAMA object as output by \code{\link{panorama_PANORAMA}}}

\item{metrics}{character. Selected metrics to plot. By default plots all metrics}

\item{title}{character. Title to be shown on top of plot}
}
\description{
Plot scatterplots of scores in PANORAMA$RES by txcoor and colored by gene
}
