% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_in_WEXAC.R
\name{panorama_PANORAMA}
\alias{panorama_PANORAMA}
\title{PANORAMA object constructor from META table}
\usage{
panorama_PANORAMA(
  META,
  genome = NULL,
  geneAnnot = NULL,
  nCores = 1,
  verb = FALSE
)
}
\arguments{
\item{META}{data.frame Experimental design}

\item{verb}{}
}
\description{
PANORAMA object constructor from META table
}
