% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{panorama_plot_confMat}
\alias{panorama_plot_confMat}
\title{Confussion matrix plot}
\usage{
panorama_plot_confMat(
  PANORAMA,
  title = "",
  pointAlpha = 0.5,
  removeOtherRNAmods = TRUE,
  rmNotModGenes = TRUE,
  blacklist = NULL
)
}
\arguments{
\item{blacklist}{}
}
\description{
Confussion matrix plot
}
