% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{percent_formatting_fr}
\alias{percent_formatting_fr}
\title{Formatting of percentages in the French style}
\usage{
percent_formatting_fr(x)
}
\arguments{
\item{x}{a number between 0 and 1}
}
\value{
a string in the form x %
}
\description{
Formatting of percentages in the French style
}
\examples{
percent_formatting_fr(.1)
}
