% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{plot_budget}
\alias{plot_budget}
\title{Bar chart by budget}
\usage{
plot_budget(table)
}
\arguments{
\item{table}{}
}
\value{
a highchart
}
\description{
Bar chart by budget
}
\examples{
plot_budget(table = table_panorama)
}
