% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plot_clusters}
\alias{plot_clusters}
\title{Plot clusters data}
\usage{
plot_clusters(G, clusters, col.names, working.dir, plot.type,
  pool.clusters = FALSE, pool.samples = FALSE, samples.to.plot = NULL,
  facet.by = "sample")
}
\arguments{
\item{G}{An \code{igraph} graph object}

\item{clusters}{A numeric vector representing the indices of the vertices of \code{G} to plot (i.e. the clusters)}

\item{col.names}{A character vector of variables to plot (these must correspond to the columns in the original clusters data)}

\item{working.dir}{The working directory. Must contain two subdirectories named \code{clusters_data} and \code{landmarks_data}. The latter
is only necessary if \code{G} contains nodes of type \code{landmark} (i.e. it represents a Scaffold map)}

\item{plot.type}{Either \code{"Boxplot"} or \code{"Scatterplot"}. The type of plot}

\item{pool.clusters}{Whether to pool the clusters data. If this is \code{FALSE} each cluster is plotted separately}

\item{pool.samples}{Whether to pool data from different samples. If this is \code{FALSE} each sample is plotted separately. Note that \code{pool.clusters}
take precedence over this (i.e. if the clusters are pooled, the samples will be pooled too, irrespective of the value of this option)}

\item{samples.to.plot}{The samples to be plotted. This option is only used if \code{pool.samples == FALSE} and \code{G} does not have 
a \code{sample} vertex property (i.e. the vertices were generated using \code{scfeatures::cluster_fcs_files_groups}, and as such they represent
multiple samples). In such cases this option is mandatory. Otherwise, if the nodes (i.e. the clusters) represent single samples
the samples to be plotted are selected from the \code{sample} vertex property of the corresponding nodes}

\item{facet.by}{In cases where multiple samples are plotted, whether the plots should be faceted by \code{"sample"} or by \code{"variable"}}
}
\value{
Returns a \code{ggplot2} plot object
}
\description{
This function plots data for individual clusters and landmarks and has different behaviours depending 
  on whether the clusters represent individual or multiple samples (i.e. wether they were generated using
  \code{scfeatures::cluster_fcs_files} or \code{scfeatures::cluster_fcs_files_groups})
}
