% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{plot_tranchebudget}
\alias{plot_tranchebudget}
\title{Pie charts of projects by budget}
\usage{
plot_tranchebudget(table)
}
\arguments{
\item{table}{}
}
\value{
a highchart
}
\description{
draw the chart by budget bucket
}
\examples{
plot_tranchebudget(table = table_panorama)

}
