% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{plot_zone}
\alias{plot_zone}
\title{Bar chart by zone}
\usage{
plot_zone(table)
}
\arguments{
\item{table}{}
}
\value{
a highchart
}
\description{
Bar chart by zone
}
\examples{
plot_zone(table = table_panorama)
}
