% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCing.R
\name{qc_plotMetricsNArate}
\alias{qc_plotMetricsNArate}
\title{Plot metrics NA rate}
\usage{
qc_plotMetricsNArate(PANORAMA)
}
\arguments{
\item{PANORAMA}{PANORAMA object}
}
\value{
ggplot
}
\description{
Plot metrics NA rate
}
