% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCing.R
\name{qc_plotPredsNApct}
\alias{qc_plotPredsNApct}
\title{Plot predictions NA rate}
\usage{
qc_plotPredsNApct(PANORAMA)
}
\arguments{
\item{PANORAMA}{PANORAMA object}
}
\value{
ggplot
}
\description{
Plot predictions NA rate
}
