% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCing.R
\name{qc_predsNApct}
\alias{qc_predsNApct}
\title{QC Predictions NA percentage}
\usage{
qc_predsNApct(PANORAMA)
}
\arguments{
\item{PANORAMA}{PANORAMA object}
}
\value{
data.table
}
\description{
QC Predictions NA percentage
}
