% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{resolve_RNAmodsByDist}
\alias{resolve_RNAmodsByDist}
\title{Resolve RNAmod multicalling by distance}
\usage{
resolve_RNAmodsByDist(
  insTable,
  PANORAMA,
  PANORAMA_training,
  models,
  blacklist_train = NULL,
  nCores = 1
)
}
\arguments{
\item{nCores}{}
}
\description{
Resolves conflicting RNAmod multicalling by calculating the geometric median
of the training sites and the distance of the conflicting sites to them per
SET (sample).
}
\details{
The space where the geometric median is calculated is in one with all the metrics
involved in estimating the conflicting RNAmods.

Additionally it generates a consensus of the calledRNAmods per position for
all the samples. And a consensus of 2-O-methylations detection confidence.
}
