% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{rm_knownRNAmods}
\alias{rm_knownRNAmods}
\title{Remove known RNAmods ('nuc' column)}
\usage{
rm_knownRNAmods(PANORAMA)
}
\arguments{
\item{PANORAMA}{PANORAMA object}
}
\value{
PANORAMA object
}
\description{
Remove known RNAmods ('nuc' column)
}
