% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RNAMod_baseNuc}
\alias{RNAMod_baseNuc}
\title{RNAmods' base nucleotides}
\format{
list
}
\usage{
RNAMod_baseNuc
}
\description{
A list containing the base nucleotides for different RNAmods used in PANORAMA
}
\keyword{datasets}
