% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RNAMod_nucRef}
\alias{RNAMod_nucRef}
\title{RNAmods' base nucleobases}
\format{
list
}
\usage{
RNAMod_nucRef
}
\description{
A list containing the base nucleobases for different RNAmods used in PANORAMA
}
\keyword{datasets}
