% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RNAmods_vec_ext}
\alias{RNAmods_vec_ext}
\title{RNAmods vector (extended)}
\format{
character
}
\usage{
RNAmods_vec_ext
}
\description{
As RNAmods_vec, but includes RNAmods that are combinations of RNAmods detected
by PANORAMA
}
\keyword{datasets}
