% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RNAmods_vec_plus}
\alias{RNAmods_vec_plus}
\title{RNAmods vector (Plus)}
\format{
character
}
\usage{
RNAmods_vec_plus
}
\description{
As RNAmods_vec_ext, but includes RNAmods that we have encountered along PANORAMA
}
\keyword{datasets}
