% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{subsetPANORAMA_byGenes}
\alias{subsetPANORAMA_byGenes}
\title{Subset PANORAMA object by genes}
\usage{
subsetPANORAMA_byGenes(PANORAMA, genes)
}
\arguments{
\item{PANORAMA}{list}

\item{genes}{character}
}
\value{
list. PANORAMA object
}
\description{
Subset PANORAMA object by genes
}
