% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{trainLinModel_allRNAmods}
\alias{trainLinModel_allRNAmods}
\title{Train all linear models}
\usage{
trainLinModel_allRNAmods(
  PANORAMA,
  nCores = 6,
  n_thresholds = 100,
  min_AUC = 0.9,
  min_MCC = 0.5,
  blacklist = NULL,
  rmGeneNoRNAmod = FALSE,
  rmRNAmod_thrSel = TRUE,
  removeOtherRNAmods = FALSE,
  rmUnmodTRNA = TRUE,
  rmOtherRNAmodKeepConf = TRUE,
  nucBias = TRUE,
  max_NArate = 0.4,
  geneInfo = NULL
)
}
\arguments{
\item{PANORAMA}{list. PANORAMA object}

\item{nCores}{integer. Number of cores to use}

\item{n_thresholds}{integer. Number of thresholds the range from the median value
of non-modified nucleotides to the median value of known modified nucleotides
is divided to test for ideal linear score threshold.}

\item{min_AUC}{numeric. Minimum AUC needed for a metric to be considered into
the linear integrated score}

\item{min_MCC}{numeric Minimum MCC for model to be output. Default = 0.7}

\item{blacklist}{data.frame. Table that includes the genes and txcoor to be
ignored from the known RNAmod nucleotides reference}

\item{rmGeneNoRNAmod}{logical. Remove genes with not RNAmods}

\item{removeOtherRNAmods}{logical. Remove other RNAmods from training step}

\item{rmOtherRNAmodKeepConf}{logical. Remove other RNAmods keeping confounding
RNAmods. Defined by \code{confRNAmods_list}.}

\item{nucBias}{logical. Only consider the values measured in the base-nucleotides
relevant to the RNAmod for training the model, e.g. only adenines for m1A}

\item{geneInfo}{}
}
\value{
list. Linear models and diagnostic plots for each trained RNAmod
}
\description{
Train all linear models
}
