% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{trainLinModel_RNAmods}
\alias{trainLinModel_RNAmods}
\title{Training linear model}
\usage{
trainLinModel_RNAmods(
  PANORAMA,
  RNAmod,
  n_thresholds = 100,
  min_AUC = 0.9,
  min_MCC = 0.5,
  blacklist = NULL,
  rmGeneNoRNAmod = FALSE,
  rmRNAmod_thrSel = TRUE,
  removeOtherRNAmods = FALSE,
  rmUnmodTRNA = TRUE,
  rmOtherRNAmodKeepConf = TRUE,
  nucBias = TRUE,
  max_NArate = 0.4,
  geneInfo = NULL
)
}
\arguments{
\item{rmRNAmod_thrSel}{logical. Removes annotated modified nucleotides for
threshold selection}

\item{removeOtherRNAmods}{logical. Removes all other RNAmods}

\item{rmOtherRNAmodKeepConf}{logical. Removes all other RNAmods but keeps confounding RNAmods}

\item{nucBias}{logical. Removes all positions for which the refSeq is not the same as the reference nucleotide}

\item{max_NArate}{logical. Maximum allowed NA rate for metric in modified sites}

\item{geneInfo}{}
}
\value{
list
}
\description{
Training linear model
}
