% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{trainLogModel_RNAmods}
\alias{trainLogModel_RNAmods}
\title{Logistic Model training using PANORAMA$CALLS object.}
\usage{
trainLogModel_RNAmods(
  PANORAMA,
  isModFun,
  RNAmod,
  thresholds,
  accNArate = TRUE,
  minMCC = 0.7
)
}
\arguments{
\item{accNArate}{logical. Account for NA rate for model selection}

\item{minMCC}{}
}
\description{
Logistic Model training using PANORAMA$CALLS object.
}
