% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yeast_PANORAMA}
\alias{yeast_PANORAMA}
\title{Yeast PANORAMA object}
\format{
list
}
\usage{
yeast_PANORAMA
}
\description{
Example PANORAMA object derived from Pan-Mod-seq data of yeast.
}
\keyword{datasets}
