% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{plot_phase}
\alias{plot_phase}
\title{Bar chart by phase}
\usage{
plot_phase(table)
}
\arguments{
\item{table}{a data frame}
}
\value{
a hichchart
}
\description{
Bar chart by phase
}
\examples{
plot_phase(table = table_panorama)
}
